Refactor: Move stats and printing logic to new modules
Extract build statistics and output formatting into dedicated modules for better organization and reusability.
