+++ src/lib.rs
@@ -3,6 +3,6 @@ pub mod commands;
 pub mod error;
 pub mod layout;
 pub mod output;
-pub mod printer;
 pub mod variant;
+pub mod visuals;
 pub mod writer;
