+++ src/visuals/mod.rs
@@ -1,5 +1,8 @@
-use crate::commands::BuildStats;
+mod formatters;
+
+use crate::commands::stats::BuildStats;
 use comfy_table::{Attribute, Cell, ContentArrangement, Table};
+use formatters::{format_address_range, format_bytes, format_efficiency};
 
 pub fn print_summary(stats: &BuildStats) {
     println!(
@@ -70,29 +73,4 @@ pub fn print_detailed(stats: &BuildStats) {
     }
 
     println!("{detail_table}");
-}
-
-fn format_bytes(bytes: usize) -> String {
-    let s = bytes.to_string();
-    let mut result = String::new();
-    for (i, c) in s.chars().rev().enumerate() {
-        if i > 0 && i % 3 == 0 {
-            result.push(',');
-        }
-        result.push(c);
-    }
-    result.chars().rev().collect::<String>() + " bytes"
-}
-
-fn format_address_range(start: u32, allocated: u32) -> String {
-    let end = start + allocated - 1;
-    format!("0x{:08X}-0x{:08X}", start, end)
-}
-
-fn format_efficiency(used: u32, allocated: u32) -> String {
-    if allocated == 0 {
-        "0.0%".to_string()
-    } else {
-        format!("{:.1}%", (used as f64 / allocated as f64) * 100.0)
-    }
 }
